Bug #11 - memswp() in qsort() not type correct. Fixed. Bug #12 - Documentation for atexit() wrong. Fixed. Bug #13 - Missing NULL include in strtox_prelim.c. Fixed. Bug #14 - s / us - unknown suffix. Fixed. Bug #15 - Wrong macro name "concat". Fixed.
git-svn-id: https://srv183.svn-repos.de/dev34/pdclib/trunk@244 bcf39385-58cc-4174-9fcf-14f50f90dd47

+++ functions/stdlib/qsort.c
@@ -19,7 +19,7 @@
 */
 
 /* Wrapper for _PDCLIB_memswp protects against multiple argument evaluation. */
-static inline void memswp( char * i, char * j, unsigned int size )
+static inline void memswp( char * i, char * j, size_t size )
 {
     _PDCLIB_memswp( i, j, size );
 }
